+++ src/lib/kotlin/slatekit-cache/build.gradle
@@ -81,10 +81,12 @@ dependencies {
     if( slatekitSetupViaMaven ) {
         compile "com.slatekit:slatekit-results:$slatekit_version"
         compile "com.slatekit:slatekit-common:$slatekit_version"
+        compile "com.slatekit:slatekit-core:$slatekit_version"
     } else {
         // */
         compile project(":slatekit-result")
         compile project(":slatekit-common")
+        compile project(":slatekit-core")
     // } //</slatekit_local>
 }
