+++ src/lib/kotlin/slatekit-cache/src/main/kotlin/slatekit/cache/CacheCommand.kt
@@ -1,4 +1,4 @@
-package slatekit.core.cache
+package slatekit.cache
 
 sealed class CacheCommand {
     abstract val action: CacheAction
