+++ src/lib/kotlin/slatekit-cache/src/main/kotlin/slatekit/cache/CacheItem.kt
@@ -11,7 +11,7 @@
  * </slate_header>
  */
 
-package slatekit.core.cache
+package slatekit.cache
 
 import slatekit.common.DateTime
 import slatekit.tracking.Tracked
