+++ src/lib/kotlin/slatekit-cache/src/main/kotlin/slatekit/cache/Management.kt
@@ -1,4 +1,4 @@
-package slatekit.core.cache
+package slatekit.cache
 
 /**
  * Core management operations on the cache.
