+++ src/lib/kotlin/slatekit-cache/src/main/kotlin/slatekit/cache/SimpleCache.kt
@@ -11,7 +11,7 @@
  * </slate_header>
  */
 
-package slatekit.core.cache
+package slatekit.cache
 
 import java.util.concurrent.ConcurrentHashMap
