--- src/lib/kotlin/slatekit-core/src/main/kotlin/slatekit/core/cache/CacheSettings.kt
@@ -1,16 +1,0 @@
-/**
- * <slate_header>
- * url: www.slatekit.com
- * git: www.github.com/code-helix/slatekit
- * org: www.codehelix.co
- * author: Kishore Reddy
- * copyright: 2016 CodeHelix Solutions Inc.
- * license: refer to website and/or github
- * about: A tool-kit, utility library and server-backend
- * mantra: Simplicity above all else
- * </slate_header>
- */
-
-package slatekit.core.cache
-
-data class CacheSettings(val size: Int)
